requirements.txt
yt-dlp
ffmpeg-python
aiogram
aiosqlite
python-dotenv
rich
loguru
pillow
aiocache
pyppeteer
aria2p
telethon